refactor: move test to injection
